Moved process for loadout optimiser into a web worker.
