+++ package.json
@@ -137,7 +137,8 @@
     "webpack-cli": "^3.0.0",
     "webpack-dev-server": "^3.8.2",
     "webpack-notifier": "^1.6.0",
-    "workbox-webpack-plugin": "^5.1.1"
+    "workbox-webpack-plugin": "^5.1.1",
+    "worker-plugin": "^4.0.3"
   },
   "dependencies": {
     "@babel/runtime": "^7.1.2",
@@ -161,6 +162,7 @@
     "ceaser-easing": "^1.0.1",
     "circuit-breaker-js": "^0.0.1",
     "clsx": "^1.0.4",
+    "comlink": "^4.3.0",
     "core-js": "^3.6.0",
     "fast-copy": "^2.0.1",
     "fast-equals": "^2.0.0",
