+++ src/app/loadout-builder/generated-sets/GeneratedSet.tsx
@@ -9,7 +9,7 @@ import { getNumValidSets, calculateTotalTier, statTier, sumEnabledStats } from '
 import { D2ManifestDefinitions } from 'app/destiny2/d2-definitions';
 import BungieImage from 'app/dim-ui/BungieImage';
 import { DestinyStatDefinition } from 'bungie-api-ts/destiny2';
-import { statHashes } from '../process';
+import { statHashes } from '../utils';
 import { t } from 'app/i18next-t';
 import styles from './GeneratedSet.m.scss';
 import { editLoadout } from 'app/loadout/LoadoutDrawer';
