+++ src/app/loadout-builder/processWorker/ProcessWorker.ts
@@ -1,0 +1,9 @@
+import { expose } from 'comlink';
+import { process } from './process';
+
+const exports = {
+  process,
+};
+export type ProcessWorker = typeof exports;
+
+expose(exports);
